refactor: move token function
